remove ability to see bribes if not game owner and add confirmation to delete game

+++ client/app/main/main.controller.js
@@ -69,9 +69,11 @@ angular.module('storyApp').controller('MainCtrl', function ($scope, $http, $loca
   };
 
   $scope.deleteGame = function(game) {
-    $http.delete('/api/games/' + game._id).then(function(){
-      $scope.getGames();
-    })
+		if(confirm("Are you sure you want to delete this game?")) {
+	    $http.delete('/api/games/' + game._id).then(function(){
+	      $scope.getGames();
+	    })
+		}
   };
 	
 	$scope.hasPermissionToPlayGame = function(gameCreator) {
